validation reduntant

+++ labelord.py
@@ -562,41 +562,40 @@ def post():
                     session = prepare_session_web(config_file)
             repos = get_repos_web(config_file, session)
             label_name = data['label']['name']
+            repo_name_payload = data['repository']['full_name']
             response_mess = ""
             if app.last_action != data["action"] or app.last_updated_label != label_name:
                 for repo in repos:
-                    label_color = data['label']['color']
-                    label_name = data['label']['name']
-                    app.last_updated_label = label_name
-                    # if not repo in app.updated_repos and repo_name_payload != repo:
-                    # app.updated_repos.append(repo)
-                    if data["action"] == "created":
-                        app.last_action = "created"
-                        header_data = {"name": label_name, "color": label_color}
-                        url = 'https://api.github.com/repos/' + repo + '/labels'
-                        response = session.post(url, json.dumps(header_data))
-                        response_mess = response_mess + repo + str(response)
-                        if response.json():
-                            response_mess = response_mess + str(response.json())
-                    if data['action'] == 'deleted':
-                        app.last_action = "deleted"
-                        url = 'https://api.github.com/repos/' + repo + '/labels/' + label_name
-                        response = session.delete(url)
-                        response_mess = response_mess + repo + url
-                        # # return str(response.json())
-                        # if response.json():
-                        #     response_mess = response_mess + str(response.json())
-                    if data['action'] == 'edited':
-                        app.last_action = "edited"
-                        new_name = label_name
-                        if "name" in data["changes"]:
-                           label_name = data["changes"]["name"]["from"]
-                        header_data = {"name": new_name, "color": label_color}
-                        url = 'https://api.github.com/repos/' + repo + '/labels/' + label_name
-                        response = session.patch(url, json.dumps(header_data))
-                        response_mess = response_mess + repo + str(response)
-                        if response.json():
-                            response_mess = response_mess + str(response.json())
+                    if repo != repo_name_payload:
+                        label_color = data['label']['color']
+                        label_name = data['label']['name']
+                        app.last_updated_label = label_name
+                        # if not repo in app.updated_repos and repo_name_payload != repo:
+                        # app.updated_repos.append(repo)
+                        if data["action"] == "created":
+                            app.last_action = "created"
+                            header_data = {"name": label_name, "color": label_color}
+                            url = 'https://api.github.com/repos/' + repo + '/labels'
+                            response = session.post(url, json.dumps(header_data))
+                            response_mess = response_mess + repo + str(response)
+                            if response.json():
+                                response_mess = response_mess + str(response.json())
+                        if data['action'] == 'deleted':
+                            app.last_action = "deleted"
+                            url = 'https://api.github.com/repos/' + repo + '/labels/' + label_name
+                            response = session.delete(url)
+                            response_mess = response_mess + repo + url
+                        if data['action'] == 'edited':
+                            app.last_action = "edited"
+                            new_name = label_name
+                            if "name" in data["changes"]:
+                                label_name = data["changes"]["name"]["from"]
+                            header_data = {"name": new_name, "color": label_color}
+                            url = 'https://api.github.com/repos/' + repo + '/labels/' + label_name
+                            response = session.patch(url, json.dumps(header_data))
+                            response_mess = response_mess + repo + str(response)
+                            if response.json():
+                                response_mess = response_mess + str(response.json())
                         # return str(url + str(header_data))
 
                 # app.updated_repos = []
